+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## [`main`](https://github.com/elastic/eui/tree/main)
 
-No public interface changes since `49.0.0`.
+- Added `readOnly` prop to `EuiMarkdownEditor` ([#5627](https://github.com/elastic/eui/pull/5627))
 
 ## [`49.0.0`](https://github.com/elastic/eui/tree/v49.0.0)